requirements.txt
h5py==3.7.0
hdbscan==0.8.28
matplotlib==3.5.3
numpy==1.23.2
pandas==1.4.3
Pillow==9.2.0
pyransac==1.1.3
scikit_image==0.19.3
scikit_learn==1.1.2
scipy==1.9.0
tensorflow==2.9.1
tqdm==4.64.0
pyransac==1.1.3
